+++ app/javascript/lca/components/header/qcHeader.jsx
@@ -1,8 +1,9 @@
 // @flow
-import React, { Fragment } from 'react'
+import React from 'react'
 import DocumentTitle from 'react-document-title'
 import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
+import { compose } from 'recompose'
 
 import { withStyles } from '@material-ui/core/styles'
 import Button from '@material-ui/core/Button'
@@ -11,6 +12,7 @@ import Typography from '@material-ui/core/Typography'
 
 import { GenericHeader } from './header.jsx'
 import LcaDrawerButton from './lcaDrawerButton.jsx'
+import QcHeaderMenu from './QcHeaderMenu.jsx'
 import { getSpecificQc, canIEditQc } from 'selectors'
 import type { fullQc } from 'utils/flow-types'
 
@@ -22,7 +24,13 @@ const styles = theme => ({
   title: {},
 })
 
-type Props = { qc: fullQc, id: number, path: string, classes: Object }
+type Props = {
+  qc: fullQc,
+  id: number,
+  path: string,
+  canIEdit: boolean,
+  classes: Object,
+}
 function QcHeader(props: Props) {
   if (props.qc == undefined) return <GenericHeader />
 
@@ -36,7 +44,7 @@ function QcHeader(props: Props) {
   }
 
   return (
-    <Fragment>
+    <>
       <DocumentTitle title={`${qc.name} | Lot-Casting Atemi`} />
 
       <Toolbar>
@@ -47,11 +55,15 @@ function QcHeader(props: Props) {
           {qc.name}
         </Typography>
 
-        <Button component={Link} to={editButtonPath} color="inherit">
-          {editing ? 'Done' : 'Edit'}
-        </Button>
+        {props.canIEdit && (
+          <Button component={Link} to={editButtonPath} color="inherit">
+            {editing ? 'Done' : 'Edit'}
+          </Button>
+        )}
+        <div className={classes.tabs} />
+        <QcHeaderMenu qc={qc} />
       </Toolbar>
-    </Fragment>
+    </>
   )
 }
 
@@ -70,4 +82,7 @@ function mapStateToProps(state, ownProps) {
   }
 }
 
-export default withStyles(styles)(connect(mapStateToProps)(QcHeader))
+export default compose(
+  withStyles(styles),
+  connect(mapStateToProps)
+)(QcHeader)